server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask==3.0.2",
    "flask-cors==4.0.0",
    "numpy>=1.24.0",
    "pillow==9.5.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "uuid==1.30",
    "werkzeug==3.0.1",
]
